refactor: move MockServer class in separate file
