+++ tests/test_get_receipt_query.py
@@ -15,7 +15,7 @@ from hiero_sdk_python.hapi.services import (
 from hiero_sdk_python.query.transaction_get_receipt_query import TransactionGetReceiptQuery
 from hiero_sdk_python.response_code import ResponseCode
 
-from tests.mock_test import mock_hedera_servers
+from tests.mock_server import mock_hedera_servers
 
 def test_transaction_get_receipt_query(transaction_id):
     """Test basic functionality of TransactionGetReceiptQuery with a mocked client."""
