+++ tests/test_topic_message_submit_transaction.py
@@ -15,7 +15,7 @@ from hiero_sdk_python.hapi.services import (
 )
 from hiero_sdk_python.response_code import ResponseCode
 
-from tests.mock_test import mock_hedera_servers
+from tests.mock_server import mock_hedera_servers
 
 @pytest.fixture
 def message():
